Update patch set 2
Patch Set 2: Code-Review-1

I think that should not be done here. The non-legacy sepolicy already labels it and this might cause a build error for double labelling the 1.2 hal for them. The automatic power hal detection should detect the one caf uses by default for your board and this should be labelled in the device/qcom/sepolicy/whateverYouUse.
But maybe we should label the 1.2 power hal also on legacy-um because some msm8998/sdm660 devices rather build 1.2 instead of the default 1.0 for this board. And on devices that don't overwrite the power hal choice it would not harm to have the 1.2 labelled even if it's not present.

Patch-set: 2
Reviewer: Gerrit User 1000480 <1000480@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Code-Review=-1